+++ resources/views/schema/_objectclasses.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
 use Yiisoft\Html\Html;
 
 ?>
-<table class="table table-striped">
+<table class="table table-striped" data-toggle="table" data-pagination="true" data-page-size="100" data-search="true">
     <thead>
     <tr>
         <th scope="col">Object Class</th>