+++ src/Asset/BootstrapTableAsset.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Asset;
+
+use Yiisoft\Assets\AssetBundle;
+
+final class BootstrapTableAsset extends AssetBundle
+{
+    public ?string $basePath = '@assets';
+    public ?string $baseUrl = '@assetsUrl';
+    public ?string $sourcePath = '@npm/bootstrap-table/dist';
+    public ?int $cssPosition = \Yiisoft\View\WebView::POSITION_BEGIN;
+
+    public array $css = [
+        'bootstrap-table.css',
+    ];
+
+    public array $js = [
+        'bootstrap-table.js',
+    ];
+}