+++ src/Asset/JQueryAsset.php
@@ -3,6 +3,7 @@
 namespace App\Asset;
 
 use Yiisoft\Assets\AssetBundle;
+use Yiisoft\View\View;
 
 class JQueryAsset extends AssetBundle
 {
@@ -11,6 +12,8 @@ class JQueryAsset extends AssetBundle
     public ?string $baseUrl = '@assetsUrl';
 
     public ?string $sourcePath = '@npm/jquery/dist';
+
+    public ?int $jsPosition = \Yiisoft\View\WebView::POSITION_HEAD;
 
     public array $js = [
         'jquery.slim.js',