+++ src/Controller/UserController.php
@@ -1,0 +1,39 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Controller;
+
+use App\Ldap\LdapService;
+use App\Service\WebControllerService;
+use Psr\Http\Message\ResponseInterface;
+use Yiisoft\Assets\AssetManager;
+use Yiisoft\Router\UrlGeneratorInterface;
+use Yiisoft\Session\Flash\FlashInterface;
+use Yiisoft\Session\SessionInterface;
+use Yiisoft\Yii\View\ViewRenderer;
+
+final class UserController
+{
+    public function __construct(public ViewRenderer          $viewRenderer,
+                                public LdapService           $ldapService,
+                                public WebControllerService  $webService,
+                                public UrlGeneratorInterface $urlGenerator,
+                                public SessionInterface      $session,
+                                public AssetManager          $assetManager,
+                                public FlashInterface        $flash
+    )
+    {
+        $this->viewRenderer = $viewRenderer
+            ->withControllerName('user')
+            ->withLayout('@views/layout/main');
+    }
+
+    public function list(WebControllerService $webService): ResponseInterface
+    {
+        return $this->viewRenderer->render('list', [
+            'urlGenerator' => $this->urlGenerator,
+            'users' => \App\Ldap\User::getAll()
+        ]);
+    }
+}